Make the stop of the neat proxy dependent on exp type
If this schedid is not a neat proxy experiemnt do not stop the proxy and firewall. 
(Try to be as resitatnt as possible if container-stop is called for a differetn schedid then the running (should not happen)

+++ files/usr/bin/container-stop.sh
@@ -20,6 +20,7 @@ if [ -f $BASEDIR/$SCHEDID.conf ]; then
   BDEXT=$(echo $CONFIG | jq -r '.basedir // empty');
   EDUROAM_IDENTITY=$(echo $CONFIG | jq -r '._eduroam.identity // empty');
   IS_VM=$(echo $CONFIG | jq -r '.vm // empty');
+  NEAT_PROXY=$(echo $CONFIG | jq -r '.neat // empty');
 fi
 if [ ! -z "$IS_INTERNAL" ]; then
   BASEDIR=/experiments/monroe${BDEXT}
@@ -85,9 +86,12 @@ if [[ ! -z "$VMIFHASH" ]]; then
 fi
 
 ## Disable NEAT proxy ###
-rm -f /etc/circle.d/60-*-neat-proxy.rules
-circle start
-
+if [[ ! -z "$NEAT_PROXY" ]]; then # If this is a experiment using the neat-proxy
+  rm -f /etc/circle.d/60-*-neat-proxy.rules
+  circle start
+  ## Stop the neta proxy container 
+  docker stop --time=10 monroe-neat-proxy
+fi
 
 sysevent -t Scheduling.Task.Stopped -k id -v $SCHEDID
